Fix warnings in unit tests

+++ tesseract_collision/test/collision_large_dataset_unit.cpp
@@ -45,7 +45,9 @@ void runTest(tesseract::DiscreteContactManagerBase& checker, bool use_convex_mes
         link_names.push_back("sphere_link_" + std::to_string(x) + std::to_string(y) + std::to_string(z));
 
         location[link_names.back()] = sphere_pose;
-        location[link_names.back()].translation() = Eigen::Vector3d(x * delta, y * delta, z * delta);
+        location[link_names.back()].translation() = Eigen::Vector3d(static_cast<double>(x) * delta,
+                                                                    static_cast<double>(y) * delta,
+                                                                    static_cast<double>(z) * delta);
         checker.addCollisionObject(link_names.back(), 0, obj3_shapes, obj3_poses, obj3_types);
       }
     }
